% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coreFunctions.R
\name{listPackRatTables}
\alias{listPackRatTables}
\title{List Supplementary Tables in locusPackRat Project}
\usage{
listPackRatTables(project_dir = ".")
}
\arguments{
\item{project_dir}{Character: Path to project directory containing .locusPackRat}
}
\value{
data.table with table information (name, link_type, link_by, n_rows, n_cols)
}
\description{
Returns information about available supplementary tables in a project
}
